backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
distro==1.9.0
Django==5.0.1
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
loguru==0.7.2
openai==1.10.0
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
pytz==2023.3.post1
replicate==0.23.0
sniffio==1.3.0
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.9.0
mysqlclient==2.2.1
django-cors-headers==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-q2==1.6.1
watchdog==3.0.0
redis==5.0.1
